Export a CMake Config file.
Modern versions of CMake support a better way of finding dependencies,
known as the Config file.  It is similar to pkg_config .pc files and is
installed alongside the package.  Then clients can just add:

   find_package(EXPAT CONFIG)
   add_executable(foo ...)
   target_link_libraries(foo EXPAT::expat)

The client target will get all the correct include paths and compile
definitions necessary to work with the expat target, automatically.